wip: migration commands for core packages
