+++ package.json
@@ -41,6 +41,7 @@
 		"migration:run": "cross-env NODE_ENV=development NODE_OPTIONS=--max_old_space_size=7000 yarn build:package:all && yarn run config:dev && yarn ts-node -r tsconfig-paths/register --project apps/api/tsconfig.app.json ./apps/api/src/migration.ts migration:run",
 		"migration:revert": "cross-env NODE_ENV=development NODE_OPTIONS=--max_old_space_size=7000 yarn build:package:all && yarn run config:dev && yarn ts-node -r tsconfig-paths/register --project apps/api/tsconfig.app.json ./apps/api/src/migration.ts migration:revert",
 		"migration:generate": "cross-env NODE_ENV=development NODE_OPTIONS=--max_old_space_size=7000 yarn build:package:all && yarn run config:dev && yarn ts-node -r tsconfig-paths/register --project apps/api/tsconfig.app.json ./apps/api/src/migration.ts migration:generate",
+		"migration:generate:core": "cross-env NODE_ENV=development NODE_OPTIONS=--max_old_space_size=4096 yarn --cwd ./packages/core migration:generate",
 		"migration:create": "cross-env NODE_ENV=development NODE_OPTIONS=--max_old_space_size=7000 yarn build:package:all && yarn run config:dev && yarn ts-node -r tsconfig-paths/register --project apps/api/tsconfig.app.json ./apps/api/src/migration.ts migration:create",
 		"seed": "cross-env NODE_ENV=development NODE_OPTIONS=--max_old_space_size=7000 yarn build:package:all && yarn run config:dev && yarn ts-node -r tsconfig-paths/register --project apps/api/tsconfig.app.json ./apps/api/src/seed.ts",
 		"seed:ever": "cross-env NODE_ENV=development NODE_OPTIONS=--max_old_space_size=7000 yarn build:package:all && yarn run config:dev && yarn ts-node -r tsconfig-paths/register --project apps/api/tsconfig.app.json ./apps/api/src/seed-ever.ts",
