+++ packages/core/package.json
@@ -26,7 +26,7 @@
 		"test": "jest",
 		"test:e2e": "jest --config ./__test__/jest.config.json",
 		"typeorm": "ts-node -r tsconfig-paths/register ./../../node_modules/typeorm/cli.js",
-		"migration:generate": "yarn run typeorm migration:generate -n",
+		"migration:generate": "yarn ts-node -r tsconfig-paths/register ./src/database/migration.ts migration:generate",
 		"build": "nest build",
 		"postbuild": "yarn run copy-files",
 		"copy-files": "copyfiles -u 1 src/**/**/*.gql src/**/**/*.hbs src/**/**/*.mjml src/**/*.csv src/**/*.json dist/",
