+++ packages/core/src/database/index.ts
@@ -1,1 +1,1 @@
-export * from './migration';
+export * from './migration-executor';
